issue #4 차트 구현 1. SaveAsImage()함수로 문제확인 : 아마 탭 또는 자식 화면과 연결하는 부분에 문제가 있어 화면에 보여주지 못하는 듯 함 --> 수정예정

+++ graduate/ChartCtrl_source/ChartCtrl.h
@@ -231,7 +231,7 @@ public:
 			The Id of the series to retrieve
 		@return The series or NULL if uSerieId is not attributed.
 	**/
-	CChartSerie* GetSerie(unsigned uSerieId) const;
+	CChartSerie* GetSerie(size_t uSerieId) const;
 	//! Removes a specific series from the chart
 	/**
 		@param uSerieId
